requirements.txt
watchdog
pytesseract
pillow